refactor 검색조건 DTO클래스로 통합
